Move client structures into entities files
