+++ azure/hostedservice/entities.go
@@ -2,7 +2,14 @@ package hostedservice
 
 import (
 	"encoding/xml"
+
+	"github.com/MSOpenTech/azure-sdk-for-go/azure"
 )
+
+//HostedServiceClient is used to manage operations on Azure Hosted Services
+type HostedServiceClient struct {
+	client *azure.Client
+}
 
 type HostedServiceDeployment struct {
 	XMLName        xml.Name `xml:"CreateHostedService"`
