+++ azure/storageservice/entities.go
@@ -2,7 +2,14 @@ package storageservice
 
 import (
 	"encoding/xml"
+
+	"github.com/MSOpenTech/azure-sdk-for-go/azure"
 )
+
+//StorageServiceClient is used to manage operations on Azure Storage
+type StorageServiceClient struct {
+	client *azure.Client
+}
 
 type StorageServiceList struct {
 	XMLName         xml.Name         `xml:"StorageServices"`
