+++ azure/virtualmachineimage/entities.go
@@ -2,7 +2,14 @@ package virtualmachineimage
 
 import (
 	"encoding/xml"
+
+	"github.com/MSOpenTech/azure-sdk-for-go/azure"
 )
+
+//ImageClient is used to manage operations on Azure Locations
+type ImageClient struct {
+	client *azure.Client
+}
 
 type ImageList struct {
 	XMLName  xml.Name  `xml:"Images"`
